feat:  Add feature to select an existing fertilizer as template for a new fertilizer
