+++ .changeset/poor-towns-appear.md
@@ -1,0 +1,5 @@
+---
+"@svenvw/fdm-app": minor
+---
+
+Add feature to select an existing fertilizer as template for a new fertilizer
